Quick and dirty iBeacon region/ranging iOS app.
Enter the UUID and hit scan and it will list the names of any matching
iBeacon regions the device has entered and any iBeacon devices that are
ranged.

The region list will typically be empty until you actually leave and
enter the region. Removing the battery from the last device and waiting
for approximately 5s is a good way to simulate this ;).